move zip serialization tests to kernel tests
