+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -163,20 +163,3 @@ tf_py_test(
         "//tensorflow/python/data/ops:dataset_ops",
     ],
 )
-
-tf_py_test(
-    name = "zip_dataset_serialization_test",
-    size = "medium",
-    srcs = ["zip_dataset_serialization_test.py"],
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-        "//third_party/py/numpy",
-    ],
-)
