+++ tensorflow/python/data/kernel_tests/BUILD
@@ -864,6 +864,7 @@ tf_py_test(
     size = "small",
     srcs = ["zip_test.py"],
     deps = [
+        ":checkpoint_test_base",
         ":test_base",
         "//tensorflow/python:array_ops",
         "//tensorflow/python:client_testlib",
